Move models into separate package, tests passing
